Tool only outputs metricRelabelings
