+++ hack/monitorgen/main.go
@@ -1,4 +1,4 @@
-// Copyright 2021 The Cluster Monitoring Operator Authors
+// Copyright 2023 The Cluster Monitoring Operator Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,47 +15,74 @@
 package main
 
 import (
-	"bytes"
 	"flag"
+	"fmt"
 	"log"
 	"os"
+	"strings"
 
-	"github.com/openshift/cluster-monitoring-operator/hack/monitorgen/internal/profiles"
-	"github.com/openshift/cluster-monitoring-operator/pkg/manifests"
+	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
+	"sigs.k8s.io/yaml"
 )
 
 var (
-	path                   *string
 	minimalMetricsFilePath *string
 )
 
 func init() {
-	path = flag.String("path", "./assets/control-plane/service-monitor-kubelet.yaml", "(Required) Path to ServiceMonitor that we want to generate collection profiles")
 	minimalMetricsFilePath = flag.String("minimal-metrics-file", "", "(Optional) Path to a file containing a list of metrics that will be kept in the minimal collection profile")
 }
 
 func main() {
+	var err error
+	var metrics []string
+
 	flag.Parse()
 
-	f, err := os.ReadFile(*path)
+	if *minimalMetricsFilePath != "" {
+		metrics, err = readFileByLine(*minimalMetricsFilePath)
+		if err != nil {
+			log.Fatalf("failed to update metrics to keep from file: %e", err)
+		}
+	}
+
+	relabelConfig := keepMetrics(metrics)
+
+	yamlData, err := yaml.Marshal(relabelConfig)
 	if err != nil {
-		log.Fatalf("failed to read asset %v: %e", *path, err)
+		log.Fatalf("failed marsahling monitor: %e", err)
 	}
 
-	m, err := manifests.NewServiceMonitor(bytes.NewReader(f))
+	fmt.Println("metricRelabelings:")
+	fmt.Println(string(yamlData))
+
+}
+
+// keepMetrics goes through the metrics in the slice metrics and joins
+// in a string with "|", them returns a relabelConfig with action "keep" and the
+// joined metrics in the regex field.
+func keepMetrics(metrics []string) []monitoringv1.RelabelConfig {
+	jointMetrics := metrics[0]
+	for i := 1; i < len(metrics); i++ {
+		jointMetrics = jointMetrics + "|" + metrics[i]
+	}
+
+	
+	return []monitoringv1.RelabelConfig{
+		{
+			Action: "keep",
+			SourceLabels: []monitoringv1.LabelName{
+				"__name__",
+			},
+			Regex: fmt.Sprintf("(%s)", jointMetrics),
+		},
+	}
+}
+
+func readFileByLine(path string) ([]string, error) {
+	f, err := os.ReadFile(path)
 	if err != nil {
-		log.Fatalf("failed to convert to service monitor: %e", err)
+		return nil, fmt.Errorf("failed to read metrics file: %e", err)
 	}
-
-	// The script is only intended to be used with resources of kind ServiceMonitor
-	if m.Kind != "ServiceMonitor" {
-		return
-	}
-
-	// Don't process ServiceMonitors that were generated by this tool as they will be overwritten
-	if val, ok := m.ObjectMeta.Annotations[manifests.CollectionProfileLabel]; ok && val != "full" {
-		return
-	}
-
-	profiles.GenerateCollectionProfileMinimal(*m, *path, *minimalMetricsFilePath)
+	return strings.Split(string(f), "\n"), nil
 }
